requirements.txt
Pillow
numpy
requests
xlsxwriter